Fix bugprone-too-small-loop-variable clang-tidy check warnings

+++ src/algorithms/PVT/libs/pvt_solution.cc
@@ -34,6 +34,7 @@
 #include "geofunctions.h"
 #include <glog/logging.h>
 #include <array>
+#include <cstddef>
 
 
 Pvt_Solution::Pvt_Solution()
@@ -271,7 +272,7 @@ void Pvt_Solution::perform_pos_averaging()
                     d_avg_latitude_d = 0.0;
                     d_avg_longitude_d = 0.0;
                     d_avg_height_m = 0.0;
-                    for (unsigned int i = 0; i < d_hist_longitude_d.size(); i++)
+                    for (size_t i = 0; i < d_hist_longitude_d.size(); i++)
                         {
                             d_avg_latitude_d = d_avg_latitude_d + d_hist_latitude_d.at(i);
                             d_avg_longitude_d = d_avg_longitude_d + d_hist_longitude_d.at(i);
